Переделан экран курса валют
