+++ Finance/Presentation/Home/Presenter/HomePresenter.swift
@@ -16,7 +16,7 @@ final class HomePresenter {
     
     private let databaseService = DatabaseService.shared
     private let networkService = NetworkService.shared
-    private let service = [ChoiceService(name: "Курс валют", img: "exchange-rate.png", vc: ExchangeRateController()),
+    private let service = [ChoiceService(name: "Курс валют", img: "exchange-rate.png", vc: ExchangeRateViewController()),
                            ChoiceService(name: "Акции", img: "stock-market.png", vc: StocksController())]
 
     private var lastTransaction = [LastTransaction]()
